+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/witness/VerkleProof.java
@@ -12,16 +12,34 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.ethereum.trie.verkle;
+package org.hyperledger.besu.ethereum.core.witness;
 
 import java.util.List;
 
 import org.apache.tuweni.bytes.Bytes;
+import org.apache.tuweni.bytes.Bytes32;
 
-public record StemStateDiff(Bytes stem, List<SuffixStateDiff> suffixDiffs) {
+// TODO Maybe create a Bytes31 in Tuweni for stems?
+public record VerkleProof(
+    List<Bytes> otherStems,
+    Bytes depthExtensionPresent,
+    List<Bytes32> commitmentsByPath,
+    Bytes32 d,
+    IPAProof ipaProof) {
 
   @Override
   public String toString() {
-    return "StemStateDiff{" + "stem=" + stem + ", suffixDiffs=" + suffixDiffs + '}';
+    return "VerkleProof{"
+        + "otherStems="
+        + otherStems
+        + ", depthExtensionPresent="
+        + depthExtensionPresent
+        + ", commitmentsByPath="
+        + commitmentsByPath
+        + ", d="
+        + d
+        + ", ipaProof="
+        + ipaProof
+        + '}';
   }
 }
