+++ plugins/rocksdb/src/test/java/org/hyperledger/besu/plugin/services/storage/rocksdb/Utils.java
@@ -86,6 +86,7 @@ public class Utils {
     return switch (dataStorageFormat) {
       case FOREST -> 1;
       case BONSAI -> 2;
+      case VERKLE -> 3;
     };
   }
 }
